services/python-api/requirements.txt
# Python API Dependencies

# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment
python-dotenv==1.0.1

# Date/Time
python-dateutil==2.8.2

# Email validation
email-validator==2.1.0.post1

# Development
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0
